create common.rs, move constant MINIMUM_TOL and functions there
